split component in different files
